requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose
pydantic
redis
razorpay
